# backend/openapi.yml
openapi: "3.0.2"
info:
  title: Rest API Docs
  version: "1.0"
servers:
  - url: http://localhost:4444/
  - url: https://ciaro360-server.vercel.app/
components:
  schemas:
    ResponseWithMsg:
      type: object
      properties:
        status:
          type: boolean
        msg:
          type: string
    IncomingProcedureCreateData:
      type: object
      properties:
        title:
          type: string
          description: "Title of the procedure"
        description:
          type: string
          description: "Description of the procedure"
        content:
          type: string
          description: "Content of the procedure"
        tags:
          $ref: "#/components/schemas/IncomingTagsData"
        policies:
          type: array
          items:
            type: string
          description: "Array of policy IDs associated with the procedure"
        projectId:
          type: string
          description: "ID of the project associated with the procedure"
    IncomingTagsData:
      type: object
      properties:
        listed:
          type: array
          items:
            type: string
          description: "Array of listed tags"
        unlisted:
          type: array
          items:
            type: string
          description: "Array of unlisted tags"
    ResponseWithMsgError:
      type: object
      properties:
        status:
          type: boolean
          example: false
        msg:
          type: string
...
    GetProcedure:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_by:
          type: string
        updated_at:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - draft
            - drafted
            - reviewed
            - approved
            - rejected
            - deleted
          default: draft
        assignees:
          type: object
          properties:
            author:
              type: array
              items:
                type: string
            reviewer:
              type: array
              items:
                type: string
            approver:
              type: array
              items:
                type: string
        versions:
          type: array
          items:
            type: string
        reminder:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        policies:
          type: array
          items:
            type: string
        controls:
          type: array
          items:
            type: string
        projects:
          type: array
          items:
            type: string
        beingModified:
          type: boolean
          default: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /file:
    post:
      tags:
        - "File-Upload Routes"
      description: "Upload Files to Server-DB"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  msg:
                    type: string
                    example: 5c7f16de348efb2b8c035bde # Example file ID
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

        401:
          description: Unauthorized
      security:
        - bearerAuth: []
  /file/{id}:
    get:
      tags:
        - "File-Upload Routes"
      description: "Retrieve File from Server-DB"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the file to retrieve
          schema:
            type: string
      responses:
        200:
          description: File retrieved successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        404:
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

        401:
          description: Unauthorized
      security:
        - bearerAuth: []

  # Templates route
...
  /policy/create/file:
    post:
      tags:
        - "Policy Routes"
      summary: "Create a policy from an uploaded file"
      description: "This endpoint creates a policy from an uploaded file."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: "The file to upload and convert."
      responses:
        '200':
          description: "File processed and policy created successfully"
          content:
            text/markdown:
              schema:
                type: string
        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  msg:
                    type: string
...
  /evidences/{id}/files:
    put:
      summary: Add files to an evidence
      tags:
        - Evidences
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: The ID of the evidence to add files to.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    description: URL of files to add to the evidence.
              required:
                - files
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []

    delete:
      summary: Remove files from the evidence
      tags:
        - Evidences
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: The ID of the evidence from which to remove files.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    description: URLs of the files to remove from the evidence.
              required:
                - files
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []

  # PROCEDURES ARE NOW PROCESS_DOCS
  # /procedure/create:
  #   post:
  #     tags:
  #       - "Procedure Routes"
  #     summary: "Create a new procedure"
  #     description: "This endpoint creates a new procedure."
  #     security:
  #       - bearerAuth: []
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: "#/components/schemas/IncomingProcedureCreateData"
  #     responses:
  #       '200':
  #         description: "Procedure created successfully"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsg"
  #       '404':
  #         description: "Resource not found"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"

  # /procedure/{id}/comment:
  #   post:
  #     tags:
  #       - "Procedure Routes"
  #     summary: "Save a comment for a procedure"
  #     description: "This endpoint saves a comment for a specific procedure."
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         schema:
  #           type: string
  #         description: "ID of the procedure"
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               images:
  #                 type: array
  #                 items:
  #                   type: string
  #                 description: "Array of image URLs"
  #               comment:
  #                 type: string
  #                 description: "Comment for the procedure"
  #     responses:
  #       '200':
  #         description: "Comment saved successfully"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsg"
  #       '404':
  #         description: "Resource not found"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"

  # /procedure/{id}/version/{name}:
  #   post:
  #     tags:
  #       - "Procedure Routes"
  #     summary: "Save a version for a procedure"
  #     description: "This endpoint saves a new version for a specific procedure."
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         schema:
  #           type: string
  #         description: "ID of the procedure"
  #       - in: path
  #         name: name
  #         required: true
  #         schema:
  #           type: string
  #         description: "Name of the version"
  #     responses:
  #       '200':
  #         description: "Version saved successfully"
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 status:
  #                   type: boolean
  #                 version:
  #                   type: string

  #       '404':
  #         description: "Resource not found"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
 
  # /procedure/{id}/reminder:
  #   post:
  #     tags:
  #       - "Procedure Routes"
  #     summary: "Set a reminder for a procedure"
  #     description: "This endpoint sets a reminder for a specific procedure."
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         schema:
  #           type: string
  #         description: "ID of the procedure"
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               next_reminder:
  #                 type: string
  #                 format: date-time
  #                 description: "Date and time of the next reminder"
  #     responses:
  #       '200':
  #         description: "Reminder set successfully"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsg"
  #       '404':
  #         description: "Resource not found"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
  
  # /procedure/{id}/tags:
  #   post:
  #     tags:
  #       - "Procedure Routes"
  #     summary: "Add tags to a procedure"
  #     description: "This endpoint adds tags to a specific procedure."
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         schema:
  #           type: string
  #         description: "ID of the procedure"
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               tags:
  #                 type: array
  #                 items:
  #                   type: object
  #                   properties:
  #                     listed:
  #                       type: array
  #                       items:
  #                         type: string
  #                         description: "Array of listed tags"
  #                     unlisted:
  #                       type: array
  #                       items:
  #                         type: string
  #                         description: "Array of unlisted tags"
  #                 description: "Array of tags to be added to the procedure"
  #     responses:
  #       '200':
  #         description: "Tags added successfully"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsg"
  #       '404':
  #         description: "Resource not found"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
  
  # /procedure/{id}/policies:
  #   post:
  #     tags:
  #       - "Procedure Routes"
  #     summary: "Map procedures to policies"
  #     description: "This endpoint maps procedures to policies."
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         schema:
  #           type: string
  #         description: "ID of the procedure"
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               procedures:
  #                 type: array
  #                 items:
  #                   type: string
  #                 description: "Array of policy IDs to map to the procedure"
  #     responses:
  #       '200':
  #         description: "Procedures mapped successfully"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsg"
  #       '404':
  #         description: "Resource not found"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
  
  # /procedure/{id}/projects:
  #   post:
  #     tags:
  #       - "Procedure Routes"
  #     summary: "Map procedure to projects"
  #     description: "This endpoint maps a procedure to projects."
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         schema:
  #           type: string
  #         description: "ID of the procedure"
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               projects:
  #                 type: array
  #                 items:
  #                   type: string
  #                 description: "Array of project IDs to map to the procedure"
  #     responses:
  #       '200':
  #         description: "Procedure mapped to projects successfully"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsg"
  #       '404':
  #         description: "Resource not found"
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"

  # /procedure/{id}/controls:
  #   post:
  #     summary: Map controls to a procedure
  #     tags:
  #       - "Procedure Routes"
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         schema:
  #           type: string
  #         description: "ID of the procedure to map controls to"
  #     requestBody:
  #         required: true
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 controls:
  #                   type: array
  #                   items:
  #                     type: string
  #                   description: "Array of controls IDs to map to the procedure"
  #     responses:
  #       '200':
  #         description: Success
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsg"
  #       '404':
  #         description: Not Found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
  #     security:
  #       - bearerAuth: []

  # /procedure:
  #   get:
  #     summary: Get all procedures
  #     tags:
  #       - "Procedure Routes"
  #     parameters:
  #       - in: query
  #         name: page
  #         schema:
  #           type: string
  #         description: Page number
  #       - in: query
  #         name: count
  #         schema:
  #           type: string
  #         description: Number of procedures per page
  #       - in: query
  #         name: field
  #         schema:
  #           type: string
  #         description: Field to sort by
  #       - in: query
  #         name: order
  #         schema:
  #           type: string
  #         description: Sorting order
  #     responses:
  #       '200':
  #         description: Success
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 status:
  #                   type: boolean
  #                   example: true
  #                 data:
  #                   type: array
  #                   items:
  #                     type: object
  #                     properties:
  #                       title:
  #                         type: string
                        
  #                       description:
  #                         type: string

  #                       beingModified:
  #                         type: string
                        
  #                       created_by:
  #                         type: string
                        
  #                       updated_at:
  #                         type: string
  #                         format: date
                          
  #                       status:
  #                         type: string
  #       '404':
  #         description: Not Found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
  #     security:
  #       - bearerAuth: []

  # /procedure/count:
  #   get:
  #     summary: Get total count of procedures
  #     tags:
  #       - "Procedure Routes"
  #     responses:
  #       '200':
  #         description: Success
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 status:
  #                   type: boolean
  #                 count:
  #                   type: number

  #       '404':
  #         description: Not Found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
  #     security:
  #       - bearerAuth: []

  # /procedure/{id}/comments:
  #   get:
  #     summary: Get comments for a procedure
  #     tags:
  #       - "Procedure Routes"
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         schema:
  #           type: string
  #         description: ID of the procedure
  #       - in: query
  #         name: page
  #         schema:
  #           type: string
  #         description: Page number
  #       - in: query
  #         name: count
  #         schema:
  #           type: string
  #         description: Number of comments per page
  #     responses:
  #       "200":
  #           description: Comments retrieved successfully
  #           content:
  #             application/json:
  #               schema:
  #                 type: object
  #                 properties:
  #                   status:
  #                     type: boolean
  #                     example: true
  #                   comment:
  #                     type: array
  #                     items:
  #                       type: object
  #                       properties:
  #                         _id:
  #                           type: string
  #                         content:
  #                           type: string
  #                         images:
  #                           type: array
  #                           items:
  #                             type: string
  #                         created_by:
  #                           type: string
  #                         created_at:
  #                           type: string
  #                           format: date-time
  #                         __v:
  #                           type: integer
  #       '404':
  #         description: Not Found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
  #     security:
  #       - bearerAuth: []

  # /procedure/{id}:
  #   get:
  #     summary: Get a procedure by ID
  #     tags:
  #       - "Procedure Routes"
  #     parameters:
  #       - in: path
  #         name: id
  #         required: true
  #         schema:
  #           type: string
  #         description: ID of the procedure
  #     responses:
  #       '200':
  #         description: Success
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #               properties:
  #                 status: 
  #                   type: boolean
  #                 procedure:
  #                   $ref: "#/components/schemas/GetProcedure"
  #       '404':
  #         description: Not Found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
  #     security:
  #       - bearerAuth: []

  # /procedure/version/{versionId}:
  #   get:
  #     summary: Get a procedure version by ID
  #     tags:
  #       - "Procedure Routes"
  #     parameters:
  #       - in: path
  #         name: versionId
  #         required: true
  #         schema:
  #           type: string
  #         description: ID of the procedure version
  #     responses:
  #       '200':
  #         description: Success
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsg"
  #       '404':
  #         description: Not Found
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/ResponseWithMsgError"
  #     security:
  #       - bearerAuth: []



  # PORJECTS
...